refactor: move stories next to components
